Move irc.js from services to interfaces
